fix: move to main tests
